% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_files.R
\docType{data}
\name{abrasion}
\alias{abrasion}
\title{Data from an experiment to investigate how weight loss due to abrasion (in g/hr) of rubber
is affected by hardness (in degrees Shore) and tensile strength (in kg/cm^2).}
\value{
A data frame with 30 rows and 3 variables.
}
\description{
Data from an experiment to investigate how weight loss due to abrasion (in g/hr) of rubber
is affected by hardness (in degrees Shore) and tensile strength (in kg/cm^2).
}
\references{
This data set was taken from Davies and Goldsmith, Statistical Methods in
Research and Production, 4th edition.
}
\keyword{datasets}
