% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_files.R
\docType{data}
\name{aircraft}
\alias{aircraft}
\title{Data on the number of 'multi-fatality crashes' in the week following 17 known cases of
'murder-suicide', together with an index of newspaper coverage.}
\value{
A list containing 2 variables with 17 rows, and a variable with 1 row.
}
\description{
Data on the number of 'multi-fatality crashes' in the week following 17 known cases of
'murder-suicide', together with an index of newspaper coverage.
}
\references{
Phillips, Airplane accident fatalities increase just after newspaper
stories about murder and suicide, Science, 201, 748-750, (1978).
}
\keyword{datasets}
