% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_files.R
\docType{data}
\name{barolo}
\alias{barolo}
\title{The prices of 75cl bottles of Barolo wines in July 2010 from the website
of a Italian wine reseller. The price does not include the delivery charge.}
\value{
A list containing a variable with 103 rows, and a variable with 1 row.
}
\description{
The prices of 75cl bottles of Barolo wines in July 2010 from the website
of a Italian wine reseller. The price does not include the delivery charge.
}
\keyword{datasets}
