% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_files.R
\docType{data}
\name{births}
\alias{births}
\title{A dataset containing the time of birth (hh:mm), sex, and birth weight (in grams) for each
of 44 babies born in one 24-hour period at a Brisbane, Australia hospital.
Also included is the number of minutes since midnight for each birth.}
\source{
\url{http://ww2.amstat.org/publications/jse/secure/v7n3/datasets.dunn.cfm}
}
\value{
A list containing 4 variables with 44 rows, and a variable with 1 row.
}
\description{
A dataset containing the time of birth (hh:mm), sex, and birth weight (in grams) for each
of 44 babies born in one 24-hour period at a Brisbane, Australia hospital.
Also included is the number of minutes since midnight for each birth.
}
\keyword{datasets}

