% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_files.R
\docType{data}
\name{coagulation}
\alias{coagulation}
\title{Data on the effect of three new diets on the blood coagulation times (in seconds)
in a set of animals.}
\value{
A data frame with 18 rows and 2 variables.
}
\description{
Data on the effect of three new diets on the blood coagulation times (in seconds)
in a set of animals.
}
\keyword{datasets}

