% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_files.R
\docType{data}
\name{crickets}
\alias{crickets}
\title{Data on the frequency of chirps made by a striped ground cricket, at various ground
temperatures.}
\value{
A data frame with 15 rows and 2 variables.
}
\description{
Data on the frequency of chirps made by a striped ground cricket, at various ground
temperatures.
}
\references{
Data collected by George W. Pierce in 1948.
}
\keyword{datasets}
