% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distns.R
\name{dbvt}
\alias{dbvt}
\alias{tcontour}
\title{The Bivariate t distribution}
\usage{
dbvt(x, y, a, b, c)

tcontour(x, y, a, b, c, p = NULL, ...)
}
\arguments{
\item{x}{vector of quantities.}

\item{y}{vector of quantities.}

\item{a}{degrees of freedom. Must be strictly positive.}

\item{b}{mean vector (2x1).}

\item{c}{covariance matrix. Must be a positive definite 2x2 matrix.}

\item{p}{probability.}

\item{...}{arguments to be passed to the plot function when plotting contours.}
}
\description{
Density (and its contours) of the bivariate t distribution with a degrees of freedom,
mean vector b and covariance matrix c.
}
\examples{
dbvt(1, 2, 10, c(1,2), diag(1,2))
beta1=seq(-3,3,len=100)
beta2=seq(-3,3,len=100)
tcontour(beta1,beta2,10,c(0,0),matrix(c(1,0.8,0.8,1),ncol=2))
}
\keyword{character}

