% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/democommands.R
\name{demoCommands}
\alias{demoCommands}
\title{Launches an editor containing the commands in a demo}
\usage{
demoCommands(demo)
}
\arguments{
\item{demo}{name of demo.}
}
\description{
Launches an editor containing the commands in a demo
}
