% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distns.R
\name{dnorminvchi}
\alias{dnorminvchi}
\alias{rnorminvchi}
\title{The Normal-InvChi distribution}
\usage{
dnorminvchi(mu, sigma, b, c, g, h)

rnorminvchi(n, b, c, g, h)
}
\arguments{
\item{mu}{vector of quantities.}

\item{sigma}{vector of quantities.}

\item{b}{mean.}

\item{c}{must be strictly positive.}

\item{g}{must be strictly positive.}

\item{h}{must be strictly positive.}

\item{n}{number of observations.}
}
\description{
Density and random generation for the normal-inv-chi distribution with parameters b, c, g and h.
}
\note{
If (mu,sigma)^T~NInvChi(b,c,g,h) then mu|sigma~N(b,sigma^2/c) and sigma~Inv-Chi(g,h). Also mu~t_(2g)(b,h/(gc)).
}
\examples{
dnorminvchi(1, 1/sqrt(2), 1, 2, 5, 3)
rnorminvchi(1, 1, 2, 5, 3)
}
\keyword{character}
