% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elicitation.R
\name{elicitgzerofn}
\alias{elicitgzerofn}
\title{Function used in the elicitation of normal-gamma prior distribution by using quantiles for mu and tau}
\usage{
elicitgzerofn(g, taup1, tauquantile1, taup2, tauquantile2)
}
\arguments{
\item{g}{must be strictly positive.}

\item{taup1}{probability.}

\item{tauquantile1}{lower taup1-th quantile for tau.}

\item{taup2}{probability.}

\item{tauquantile2}{lower taup2-th quantile for tau.}
}
\value{
value
}
\description{
Function used in the elicitation of normal-gamma prior distribution by using quantiles for mu and tau
}
\keyword{character}

