% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_files.R
\docType{data}
\name{exposure}
\alias{exposure}
\title{The chlorobenzene exposure concentrations (15 min TWAs) in ppm for a
chemical worker during a full workshift for 30 consecutive days.}
\value{
A list containing a variable with 31 rows, and a variable with 1 row.
}
\description{
The chlorobenzene exposure concentrations (15 min TWAs) in ppm for a
chemical worker during a full workshift for 30 consecutive days.
}
\keyword{datasets}
