% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibbs.R
\name{gibbsReffects}
\alias{gibbsReffects}
\title{Gibbs sampler for a one-way normal random effects model with a semi-conjugate prior}
\usage{
gibbsReffects(N, initial, priorparam, m, n, ybar, s)
}
\arguments{
\item{N}{length of MCMC chain.}

\item{initial}{starting values for the population mean and the precision components.}

\item{priorparam}{prior parameters a,b,c,d,e,f.}

\item{m}{number of treatments.}

\item{n}{vector containing the number of observations on each treatment.}

\item{ybar}{vector containing the mean of observations on each treatment.}

\item{s}{vector containing the standard deviation of observations on each treatment.}
}
\description{
Simulates realisations from the posterior distribution for the population mean, random effect means
and precision components in a one-way normal random effects model with a semi-conjugate prior.
}
\examples{
data(contamination)
n=tapply(contamination$acc,contamination$keyboard,length)
ybar=tapply(contamination$acc,contamination$keyboard,mean)
s=sqrt(tapply(contamination$acc,contamination$keyboard,var)*(n-1)/n)
mcmcAnalysis(gibbsReffects(N=100,initial=c(200,2e-5,1),
 priorparam=c(200,0.1,0.1,0.1,0.1,0.1),m=10,n=n,ybar=ybar,s=s))
}

