% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_files.R
\docType{data}
\name{glassfibre}
\alias{glassfibre}
\title{The breaking strengths of 63 glass fibres of length 1.5cm.}
\value{
A list containing a variable with 63 rows, and a variable with 1 row.
}
\description{
The breaking strengths of 63 glass fibres of length 1.5cm.
}
\keyword{datasets}
