% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_files.R
\docType{data}
\name{granite}
\alias{granite}
\title{A random sample of granite gravestones was taken from a cemetery in Savannah,
a city along the southeast coast of the USA (coastal) and a cemetery in
Macon, Georgia, a town located 170 miles inland from Savannah (inland).
The dataset consists of the age of each gravestone (the difference between
the current year and the year the person died), the depth of degradation (in
mm) as determined by the lettering alteration method, and the location of the
gravestone where the value of location is 0 for coastal and 1 for inland.}
\source{
T.C. Meierdin, Marble weathering rates: a transect of the United States.
Physical Geography, 2, pp. 1-18, (1981)
}
\value{
A list containing 3 variables with 25 rows, and a variable with 1 row.
}
\description{
A random sample of granite gravestones was taken from a cemetery in Savannah,
a city along the southeast coast of the USA (coastal) and a cemetery in
Macon, Georgia, a town located 170 miles inland from Savannah (inland).
The dataset consists of the age of each gravestone (the difference between
the current year and the year the person died), the depth of degradation (in
mm) as determined by the lettering alteration method, and the location of the
gravestone where the value of location is 0 for coastal and 1 for inland.
}
\keyword{datasets}

