% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdi.R
\name{hdiInvchi}
\alias{hdiInvchi}
\title{The highest density interval (HDI) for a Inv-Chi(a,b) distribution}
\usage{
hdiInvchi(p = 0.95, a, b)
}
\arguments{
\item{p}{confidence level for the HDI.}

\item{a}{a positive number.}

\item{b}{a positive number.}
}
\value{
Vector containing the limits of the HDI.
}
\description{
The highest density interval (HDI) for a Inv-Chi(a,b) distribution
}
\examples{
hdiInvchi(0.9, 2, 2)
}
\keyword{character}

