% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_files.R
\docType{data}
\name{hepatitis}
\alias{hepatitis}
\title{Data on the logarithm of ornithine carbonyltransferase (a liver enzyme) in a random
sample of patients suffering from acute viral hepatitis.}
\value{
A data frame with 30 rows and 1 variable.
}
\description{
Data on the logarithm of ornithine carbonyltransferase (a liver enzyme) in a random
sample of patients suffering from acute viral hepatitis.
}
\keyword{datasets}
