% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_files.R
\docType{data}
\name{IBMprices}
\alias{IBMprices}
\title{Data consisting of 160 consecutive daily IBM common stock closing prices.}
\value{
A list containing a variable with 160 rows, and a variable with 1 row.
}
\description{
Data consisting of 160 consecutive daily IBM common stock closing prices.
}
\references{
Box and Jenkins, Time Series Analysis, 1976.
}
\keyword{datasets}

