% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_files.R
\docType{data}
\name{insulation}
\alias{insulation}
\title{Data on weekly gas consumption (in thousands of cubic feet) at 30 houses over 56 weeks.
The data includes the outside temperature (in degrees C), and the presence of insulation
(z=1: no insulation, z=2: insulation present).}
\value{
A data frame with 56 rows and 32 variables.
}
\description{
Data on weekly gas consumption (in thousands of cubic feet) at 30 houses over 56 weeks.
The data includes the outside temperature (in degrees C), and the presence of insulation
(z=1: no insulation, z=2: insulation present).
}
\keyword{datasets}

