% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_files.R
\docType{data}
\name{leukaemia}
\alias{leukaemia}
\title{Data on 33 leukaemia patients, classified by AG Group (1: AG positive, 2: AG negative).
For each patient white blood cell count and survival time (in weeks) is also provided.}
\value{
A list containing 2 variables with 33 rows, and a variable with 1 row.
}
\description{
Data on 33 leukaemia patients, classified by AG Group (1: AG positive, 2: AG negative).
For each patient white blood cell count and survival time (in weeks) is also provided.
}
\references{
Feigl and Zelen, Estimation of exponential survival probabilities with
concomitant information, Biometrics, 21, 826-38 (1965).
}
\keyword{datasets}
