% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_files.R
\docType{data}
\name{lungs}
\alias{lungs}
\title{Data on 50 patients suffering from a chronic lung disease. Included is 30 measurements
of lung function (y1 to y30), along with patient age (in years), sex (0: female, 1:male),
and smoking status (0:non-smoker, 1:smoker).}
\value{
A data frame with 50 rows and 33 variables.
}
\description{
Data on 50 patients suffering from a chronic lung disease. Included is 30 measurements
of lung function (y1 to y30), along with patient age (in years), sex (0: female, 1:male),
and smoking status (0:non-smoker, 1:smoker).
}
\keyword{datasets}
