% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_files.R
\docType{data}
\name{malcolm}
\alias{malcolm}
\title{Data collected by Malcolm on the height and shoe size of a random sample of students}
\value{
A data frame with 52 rows and 2 variables.
}
\description{
Data collected by Malcolm on the height and shoe size of a random sample of students
}
\keyword{datasets}
