% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{mcmcAnalysis}
\alias{mcmcAnalysis}
\title{Summarise and plot tabular MCMC output}
\usage{
mcmcAnalysis(mat, rows = 4, lag.max = 100, bins = 30, show = TRUE,
  plot = TRUE)
}
\arguments{
\item{mat}{matrix of MCMC output, where the columns represent variables and the rows represent iterations.}

\item{rows}{number of variables to plot per page on the graphics device.}

\item{lag.max}{maximum lag for the ACF plots.}

\item{bins}{approximate number of bins to use for the histograms.}

\item{show}{a logical. If TRUE, will display numerical summaries on the R console.}

\item{plot}{a logical. If TRUE, will plot graphical summaries on the default graphics device.}
}
\description{
This function summarises and plots tabular MCMC output such as that generated by the function normgibbs.
}
\examples{
posterior=gibbsNormal(N=1000,initial=c(10,0.25),priorparam=c(10,1/100,3,12),n=100,xbar=15,s=4.5)
mcmcAnalysis(posterior,rows=2,bins=10)
}
\references{
A version of mcmcSummary in Darren Wilkinson's package smfsb.
}
