% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{mcmcBoxplot}
\alias{mcmcBoxplot}
\title{Box plots from MCMC output}
\usage{
mcmcBoxplot(input, start = 1, end = NULL)
}
\arguments{
\item{input}{matrix containing MCMC realisations.}

\item{start}{take realisations starting at start.}

\item{end}{and ending at end.}
}
\description{
Box plots from MCMC output
}

