% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{mcmcCi}
\alias{mcmcCi}
\title{Equi-tailed confidence intervals from MCMC output}
\usage{
mcmcCi(input, level = 0.95)
}
\arguments{
\item{input}{matrix containing MCMC realisations.}

\item{level}{confidence level required.}
}
\description{
Equi-tailed confidence intervals from MCMC output
}

