% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_files.R
\docType{data}
\name{moths}
\alias{moths}
\title{The number of spruce moths found in traps after 48 hours by location
of trap in tree (top branches, middle branches, lower branches, ground) and type
of lure in the trap (scent, sugar, chemical).}
\value{
A list containing 3 variables with 60 rows, and a variable with 1 row.
}
\description{
The number of spruce moths found in traps after 48 hours by location
of trap in tree (top branches, middle branches, lower branches, ground) and type
of lure in the trap (scent, sugar, chemical).
}
\keyword{datasets}
