% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_files.R
\docType{data}
\name{oldfaithful}
\alias{oldfaithful}
\title{Data on old faithful}
\value{
A data frame with 299 rows with 1 variable.

A data frame with 299 rows with 1 variable.
}
\description{
Data on old faithful

Data on old faithful
}
\keyword{datasets}

