% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_files.R
\docType{data}
\name{protein}
\alias{protein}
\title{Data from an experiment to calibrate a near infrared reflectance instrument. Included is
the protein content (\%) as determined by a standard method (y1 to y30), and measurements
of reflectance at three different wavelengths (x1 to x3).}
\value{
A data frame with 24 rows and 33 variables.
}
\description{
Data from an experiment to calibrate a near infrared reflectance instrument. Included is
the protein content (\%) as determined by a standard method (y1 to y30), and measurements
of reflectance at three different wavelengths (x1 to x3).
}
\keyword{datasets}
