% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_files.R
\docType{data}
\name{rats}
\alias{rats}
\title{Data on the weights of a set of thirty young rats, measured over time (in days).}
\value{
A list with 1 matrix and 1 variable.
}
\description{
Data on the weights of a set of thirty young rats, measured over time (in days).
}
\references{
Gelfand, Hills, Racine-Poon, and Smith, Illustration of Bayesian inference
 in normal data models using Gibbs sampling, Journal of the American Statistical
 Association, 85, 972-85, (1990).
}
\keyword{datasets}
