% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_files.R
\docType{data}
\name{sales}
\alias{sales}
\title{Data on the number of orders received by a small company for one of its products over time.
The data consists of 30 sets of 100 consecutive weeks.}
\value{
A list containing 30 variables with 100 rows, and a variable with 1 row.
}
\description{
Data on the number of orders received by a small company for one of its products over time.
The data consists of 30 sets of 100 consecutive weeks.
}
\keyword{datasets}

