% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_files.R
\docType{data}
\name{shoots}
\alias{shoots}
\title{The number of roots produced by 270 micropropagated shoots of the
apple cultivar Trajan. During the rooting period, all shoots
were maintained under identical conditions but the shoots themselves
were cultured on media containing different concentrations of
cytokinin BAP, in growth cabinets with an 8 or 16 hour photoperiod.
Thus each experimental condition is decribed by two factors: BAP
concentration (2.2, 4.4, 8.8 and 17.6 $\\mu M$) and photoperiod (8
and 16 hours).}
\value{
A list containing 3 variables with 270 rows.
}
\description{
The number of roots produced by 270 micropropagated shoots of the
apple cultivar Trajan. During the rooting period, all shoots
were maintained under identical conditions but the shoots themselves
were cultured on media containing different concentrations of
cytokinin BAP, in growth cabinets with an 8 or 16 hour photoperiod.
Thus each experimental condition is decribed by two factors: BAP
concentration (2.2, 4.4, 8.8 and 17.6 $\\mu M$) and photoperiod (8
and 16 hours).
}
\keyword{datasets}
