% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_files.R
\docType{data}
\name{sideeffects}
\alias{sideeffects}
\title{Data showing the number of patients who were given different doses of a
certain drug, along with the number of patients who exhibited a
particular side effect.}
\value{
A data frame with 7 rows with 3 variables.
}
\description{
Data showing the number of patients who were given different doses of a
certain drug, along with the number of patients who exhibited a
particular side effect.
}
\keyword{datasets}

