% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_files.R
\docType{data}
\name{tonsils}
\alias{tonsils}
\title{A dataset containing the results of a study of the tonsil size (Normal, Large
or Very Large) of 1398 school children, along with whether or not they were
carriers of the bacterium Streptococcus pyogenes.}
\source{
W. J. Krzanowski, Principles of Multivariate Analysis, Oxford: Oxford
University Press, 1988.
}
\value{
A list containing 1 variable with 6 rows.
}
\description{
A dataset containing the results of a study of the tonsil size (Normal, Large
or Very Large) of 1398 school children, along with whether or not they were
carriers of the bacterium Streptococcus pyogenes.
}
\keyword{datasets}
