% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_files.R
\docType{data}
\name{oldfaithful}
\alias{oldfaithful}
\title{Data on successive waiting times (in minutes) between eruptions of the "Old Faithful"
geyser in Yellowstone National Park, Wyoming, USA.}
\value{
A list containing a variable with 299 rows, and a variable with 1 row.
}
\description{
Data on successive waiting times (in minutes) between eruptions of the "Old Faithful"
geyser in Yellowstone National Park, Wyoming, USA.
}
\references{
Azzalini and Bowman, A look at some data on the Old Faithful geyser,
Journal of the Royal Statistical Society, Series C, 39, 357-366, (1990).
}
\keyword{datasets}
